refactor: split nostr/events into dedup, admission, and processor
Break the 4k-line events module into focused submodules while keeping the
public API unchanged. Event processor tests move with the implementation.
